Rename pending to payment.
