+++ js/views/modals/purchase/Payment.js
